Version 1
A dashboard created using a shiny app created which calculates portfolio returns of 7 stocks and a graph which plots returns on the x-axis and the count on the y-axis. Different stocks can be considered by changing the ticker symbol. The weights can be adjusted as well (totaling to 100). Starting date and rebalancing can also be manipulated.